refactor: Adjust GTK keyboard handling for new KeyboardManager
