+++ src/Uno.UI.Runtime.Skia.Gtk/Extensions/GtkCoreWindowExtension.cs
@@ -20,11 +20,7 @@ namespace Uno.UI.Runtime.Skia.Gtk
 		public GtkCoreWindowExtension(object owner)
 		{
 			_owner = (CoreWindow)owner;
-
-			InitializeKeyboard();
 		}
-
-		partial void InitializeKeyboard();
 
 		internal static Fixed? GetOverlayLayer(XamlRoot xamlRoot) =>
 			GtkManager.XamlRootMap.GetHostForRoot(xamlRoot)?.NativeOverlayLayer;
@@ -119,24 +115,6 @@ namespace Uno.UI.Runtime.Skia.Gtk
 			}
 
 			return new(0, 0);
-		}
-
-		public static VirtualKeyModifiers GetKeyModifiers(Gdk.ModifierType state)
-		{
-			var modifiers = VirtualKeyModifiers.None;
-			if (state.HasFlag(Gdk.ModifierType.ShiftMask))
-			{
-				modifiers |= VirtualKeyModifiers.Shift;
-			}
-			if (state.HasFlag(Gdk.ModifierType.ControlMask))
-			{
-				modifiers |= VirtualKeyModifiers.Control;
-			}
-			if (state.HasFlag(Gdk.ModifierType.Mod1Mask))
-			{
-				modifiers |= VirtualKeyModifiers.Menu;
-			}
-			return modifiers;
 		}
 	}
 }
